Add `HttpStatus.isContentAlwaysEmpty()`
Motivation:

It is often useful to tell if the response content will always be empty
from an HTTP status code.

Modifications:

- Move `ArmeriaHttpUtil.isContentAlwaysEmpty()` to `HttpStatus`.

Result:

- `isContentAlwaysEmpty()` is now part of the public API.
